+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Development
 
+### Changed
+
+- operator selection for boolean filter attributes restricted to `EQ`/`NE`
+
 ### Added
 
 - automatic axis units (ms, bps, bytes, percentunit) inferred from the dimension or metric being aggregated
+- true/false dropdown for boolean filter attributes
+- helpful placeholder for the `IN` filter operator
 
 ## 1.6.0